# specs/001-esat-skills-portfolio/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Livret Numérique ESAT API
  description: API REST pour l'application de parcours de compétences ESAT
  version: 1.0.0
  contact:
    name: ESAT Support
servers:
  - url: http://localhost:3001/api
    description: Développement local
  - url: https://api.esat-livret.local/api
    description: Production
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Liste des utilisateurs
      security:
        - cookieAuth: []
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Créer un utilisateur
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Username déjà utilisé
  /users/{userId}:
    get:
      tags: [Users]
      summary: Détail d'un utilisateur
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Détail utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Users]
      summary: Modifier un utilisateur
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Utilisateur modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}/portfolio:
    get:
      tags: [Users]
      summary: Parcours complet d'un utilisateur (lecture seule)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Parcours complet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPortfolio'

  # ============ EVENTS ============
...
  /validations:
    get:
      tags: [Validations]
      summary: Liste des compétences en attente de validation
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Validations en attente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingValidation'
        '403':
          $ref: '#/components/responses/Forbidden'
  /validations/{userSkillId}/approve:
    post:
      tags: [Validations]
      summary: Valider une compétence
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/userSkillId'
      responses:
        '200':
          description: Compétence validée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSkill'
  /validations/{userSkillId}/reject:
    post:
      tags: [Validations]
      summary: Refuser une compétence
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/userSkillId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectValidationRequest'
      responses:
        '200':
          description: Compétence refusée

  # ============ DOCUMENTS ============
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    userSkillId:
      name: userSkillId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Forbidden:
      description: Accès interdit (rôle insuffisant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Role:
      type: string
      enum: [USER, ADMIN]
    SkillStatus:
      type: string
      enum: [IN_PROGRESS, PENDING_VALIDATION, ACQUIRED, REJECTED]
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        avatarLevel:
          type: integer
          minimum: 1
          maximum: 5
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            skillsCount:
              type: object
              properties:
                acquired:
                  type: integer
                inProgress:
                  type: integer
                pending:
                  type: integer
            eventsCount:
              type: integer
    CreateUserRequest:
      type: object
      required: [username, password, firstName, lastName, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        role:
          $ref: '#/components/schemas/Role'
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        isActive:
          type: boolean
        password:
          type: string
          minLength: 6
    UserPortfolio:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        skills:
          $ref: '#/components/schemas/MySkillsSummary'
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        title:
          type: string
          maxLength: 100
        photoUrl:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        iconName:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
...
    UserSkill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        skillId:
          type: string
          format: uuid
        skill:
          $ref: '#/components/schemas/Skill'
        status:
          $ref: '#/components/schemas/SkillStatus'
        requestedAt:
          type: string
          format: date-time
        validatedAt:
          type: string
          format: date-time
          nullable: true
        rejectionReason:
          type: string
          nullable: true
    MySkillsSummary:
      type: object
      properties:
        avatarLevel:
          type: integer
        summary:
          type: object
          properties:
            acquired:
              type: integer
            inProgress:
              type: integer
            pendingValidation:
              type: integer
            rejected:
              type: integer
        skills:
          type: array
          items:
            $ref: '#/components/schemas/UserSkill'
...
    PendingValidation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        skill:
          $ref: '#/components/schemas/Skill'
        requestedAt:
          type: string
          format: date-time
    RejectValidationRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          maxLength: 500
...